e2e: update logger path and persist logs (#7792)
### Summary

This updates the logger to support dynamically updating its file path.
The log files are renamed in the app fixture after the app shuts down,
so this ensures that if a test hangs (for any reason), logs continue to
be written to the correct file.

Also uploading `test-logs` dir in CI. Not sure we need this, I can put
it behind a flag so we can toggle. Thoughts? It is currently disabled.

### QA Notes

Verified using the large-python-notebook test: by toggling the afterAll
hook, I confirmed that trailing logs (which previously failed to write)
now correctly appear in the updated log file.

<img width="1815" alt="Screenshot 2025-05-21 at 11 42 00 AM"
src="https://github.com/user-attachments/assets/e5e9b32c-919c-4e7a-9ad3-d2e89090623e"
/>

+++ test/e2e/infra/logger.ts
@@ -3,12 +3,14 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { appendFileSync, writeFileSync } from 'fs';
+import { appendFileSync, writeFileSync, existsSync } from 'fs';
 import { format } from 'util';
 import { EOL } from 'os';
+import path from 'path';
 
 export interface Logger {
 	log(message: string, ...args: any[]): void;
+	setPath?(dir: string, filename?: string): void;
 }
 
 export class ConsoleLogger implements Logger {
@@ -19,9 +21,22 @@ export class ConsoleLogger implements Logger {
 }
 
 export class FileLogger implements Logger {
+	private path: string;
 
-	constructor(private path: string) {
-		writeFileSync(path, '');
+	constructor(initialPath: string) {
+		this.path = initialPath;
+		this.ensureFileExists(this.path);
+	}
+
+	setPath(dir: string, filename = 'e2e-test-runner.log'): void {
+		this.path = path.join(dir, filename);
+		this.ensureFileExists(this.path);
+	}
+
+	private ensureFileExists(path: string): void {
+		if (!existsSync(path)) {
+			writeFileSync(path, '');
+		}
 	}
 
 	log(message: string, ...args: any[]): void {
@@ -38,6 +53,14 @@ export class FileLogger implements Logger {
 export class MultiLogger implements Logger {
 
 	constructor(private loggers: Logger[]) { }
+
+	setPath(dir: string, filename = 'e2e-test-runner.log'): void {
+		for (const logger of this.loggers) {
+			if (logger.setPath) {
+				logger.setPath(dir, filename);
+			}
+		}
+	}
 
 	log(message: string, ...args: any[]): void {
 		for (const logger of this.loggers) {
